+++ packages/shared/src/components/enhancers/index.ts
@@ -1,4 +1,7 @@
-export * from "./ContentEnhancer";
+export * from "./ContentStyleEnhancer";
+export * from "./KeyboardShortcuts";
+export * from "./DualViewManager";
+export * from "./TextareaCounter";
+export * from "./TextareaResizer";
 export * from "./CourseListEnhancer";
 export * from "./SidebarEnhancer";
-export * from "./TextareaEnhancer";
