+++ packages/userscript-base/src/index.ts
@@ -1,5 +1,7 @@
 // Re-export all components and utilities
 export { MikotoApp } from "./components/MikotoApp";
+export { MikotoAppLite } from "./components/MikotoAppLite";
 export { createStorageManager } from "./utils/storage";
 export { GMStorageAdapter } from "./utils/storageAdapter";
 export { initializeMikoto } from "./main";
+export { initializeMikotoLite } from "./mainLite";
